+++ Mesmerist/Class/Mesmerist/Archetypes/VexingDaredevil/SloppyDefense.cs
@@ -52,10 +52,15 @@ namespace Mesmerist.Class.Mesmerist.Archetypes.VexingDaredevil
                 .AddTargetAttackWithWeaponTrigger(
                     onlyHit: true,
                     actionsOnAttacker: ActionsBuilder.New().Conditional(
-                        ConditionsBuilder.New().Add<ContextConditionInitiatorHasFact>(c =>
-                        {
-                            c.FactToCheck = BlueprintTool.GetRef<BlueprintUnitFactReference>(Guids.SloppyDefense);
-                        }),
+                        ConditionsBuilder.New()
+                            .Add<ContextConditionInitiatorHasFact>(c =>
+                            {
+                                c.FactToCheck = BlueprintTool.GetRef<BlueprintUnitFactReference>(Guids.SloppyDefense);
+                            })
+                            .Add<ContextConditionOwnerHasBuffFromCaster>(c =>
+                            {
+                                c.Buff = BlueprintTool.GetRef<BlueprintBuffReference>(Guids.FeintedBuff);
+                            }),
                         ifTrue: ActionsBuilder.New().ApplyBuff(
                             Guids.SloppyDefenseBuffEffect,
                             ContextDuration.Fixed(1, DurationRate.Rounds))))