+++ Mesmerist/Class/Mesmerist/Archetypes/VexingDaredevil/SurpriseStrike.cs
@@ -46,10 +46,15 @@ namespace Mesmerist.Class.Mesmerist.Archetypes.VexingDaredevil
                 .AddTargetAttackWithWeaponTrigger(
                     onlyHit: true,
                     actionOnSelf: ActionsBuilder.New().Conditional(
-                        ConditionsBuilder.New().Add<ContextConditionInitiatorHasFact>(c =>
-                        {
-                            c.FactToCheck = BlueprintTool.GetRef<BlueprintUnitFactReference>(Guids.SurpriseStrike);
-                        }),
+                        ConditionsBuilder.New()
+                            .Add<ContextConditionInitiatorHasFact>(c =>
+                            {
+                                c.FactToCheck = BlueprintTool.GetRef<BlueprintUnitFactReference>(Guids.SurpriseStrike);
+                            })
+                            .Add<ContextConditionOwnerHasBuffFromCaster>(c =>
+                            {
+                                c.Buff = BlueprintTool.GetRef<BlueprintBuffReference>(Guids.FeintedBuff);
+                            }),
                         ifTrue: ActionsBuilder.New().DealDamage(
                             new DamageTypeDescription() { Type = DamageType.Direct },
                             new ContextDiceValue()